% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R
\name{add_history}
\alias{add_history}
\title{Add or modify history header of an ODF}
\usage{
add_history(odfstruct, section, comment)
}
\arguments{
\item{odfstruct}{an ODF list structure, created with gen_odfstruct or read
with read_odf}

\item{section}{The history section you wish to add or modify (each
creation_date line starts a new section)}

\item{comment}{the comment you wish to insert. note: if modifying an existing
line this comment will NOT append existing comments but overwrite them}
}
\value{
an ODF list object with modified history header
}
\description{
Add or modify history header of an ODF
}
