% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/add_parameter.R, R/oce2odf.R
\name{add_parameter}
\alias{add_parameter}
\alias{add_parameter}
\alias{add_parameter}
\title{add a parameter to an existing ODF structure array (b)
reads in variable, add a parameter header, cal header and column to the data portion of the ODF}
\usage{
add_parameter(b, data, VARNAME, cal = FALSE)

add_parameter(b, data, VARNAME, cal = FALSE)

add_parameter(b, data, VARNAME, cal = FALSE)
}
\arguments{
\item{b}{odf structure}

\item{VARNAME}{name of parameter (in object data is being pulled from)}

\item{cal}{TRUE/ FALSE whether or not to produce a polynomial cal header for same parameter code}

\item{obj}{oce/ odf (?) object to  pull data and metadata from}

\item{b}{odf structure}

\item{obj}{oce/ odf (?) object to  pull data and metadata from}

\item{VARNAME}{name of parameter (in object data is being pulled from)}

\item{cal}{TRUE/ FALSE whether or not to produce a polynomial cal header for same parameter code}

\item{b}{odf structure}

\item{obj}{oce/ odf (?) object to  pull data and metadata from}

\item{VARNAME}{name of parameter (in object data is being pulled from)}

\item{cal}{TRUE/ FALSE whether or not to produce a polynomial cal header for same parameter code}
}
\description{
parameter header has: type, name, units, code, null_value, print_field_width,
print_decimal_places, angle_of_Section magnetic_Variation, depth,
minimum_value, maximum_value, number_valid, number_null

parameter header has: type, name, units, code, null_value, print_field_width,
print_decimal_places, angle_of_Section magnetic_Variation, depth,
minimum_value, maximum_value, number_valid, number_null

parameter header has: type, name, units, code, null_value, print_field_width,
print_decimal_places, angle_of_Section magnetic_Variation, depth,
minimum_value, maximum_value, number_valid, number_null
}
\details{
polynomial_cal_header has, parameter_name, calibration_date, application_date,
number_coefficients, coefficients

data columns should be added in order of parameter headers

polynomial_cal_header has, parameter_name, calibration_date, application_date,
number_coefficients, coefficients

data columns should be added in order of parameter headers

polynomial_cal_header has, parameter_name, calibration_date, application_date,
number_coefficients, coefficients

data columns should be added in order of parameter headers
}
