% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/read_odf.R
\name{addToPram}
\alias{addToPram}
\alias{addToPram}
\title{addToPram}
\usage{
addToPram(pram, name, val)

addToPram(pram, name, val)
}
\arguments{
\item{pram}{- Null or the existing list of parameters}

\item{name}{- the name of the sub-parameter list to add the value to}

\item{val}{- the value to add the the parameter ist.}

\item{pram}{- Null or the existing list of parameters}

\item{name}{- the name of the sub-parameter list to add the value to}

\item{val}{- the value to add the the parameter ist.}
}
\description{
Description:
	Used to create and add to a list, parameters using the same name are added
to a list of other parameters using the same name. The list is then returned.

Description:
	Used to create and add to a list, parameters using the same name are added
to a list of other parameters using the same name. The list is then returned.
}
