% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/oce2odf.R
\name{as.gf3}
\alias{as.gf3}
\alias{as.gf3}
\title{Translate to GF3 codes}
\usage{
as.gf3(VARNAME)

as.gf3(VARNAME)
}
\arguments{
\item{VARNAME}{the common name for known variables or the GF3 param code if known}

\item{VARNAME}{the common name for known variables or the GF3 param code if known}
}
\value{
list with GF3 code, def, width, prec and units

list with GF3 code, def, width, prec and units
}
\description{
Translate to GF3 codes

Translate to GF3 codes
}
