% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/odfedit.R
\name{batchedit}
\alias{batchedit}
\alias{batchedit}
\title{Batch Edit ODF files}
\usage{
batchedit(odffiles, param, value)

batchedit(odffiles, param, value)
}
\arguments{
\item{odffiles}{a list of odf file names which are to be edited}

\item{param}{The parameter you wish to edit}

\item{value}{The value you wish to insert into ODF files}

\item{odffiles}{a list of odf file names which are to be edited}

\item{param}{The parameter you wish to edit}

\item{value}{The value you wish to insert into ODF files}
}
\value{
edited ODF files

edited ODF files
}
\description{
This function can be used to edit a series of ODF files at a time, if a batch
of ODF files were produced with an incorrect parameter or value

This function can be used to edit a series of ODF files at a time, if a batch
of ODF files were produced with an incorrect parameter or value
}
\examples{
see editParam example

batchedit(odffiles = list.files(path = '.', pattern = '*.ODF'), param = 'ORGANIZATION', value = 'DFO SABS')


see editParam example

batchedit(odffiles = list.files(path = '.', pattern = '*.ODF'), param = 'ORGANIZATION', value = 'DFO SABS')


}
