% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfname.R
\name{bfname}
\alias{bfname}
\title{Create file name}
\usage{
bfname(odf_obj)
}
\arguments{
\item{odf_obj}{a list ODF object}
}
\value{
A standardized name for an ODF file based on metadata parameters
}
\description{
Creates standardized ODF file name for any data type
}
\author{
R. Pettipas
}
