% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/oce2odf.R
\name{binExport}
\alias{binExport}
\alias{binExport}
\title{Select bin export}
\usage{
binExport(obj, bins)

binExport(obj, bins)
}
\arguments{
\item{obj}{an ODF structure object with multiple depth bins (produced from oce2odf)}

\item{bins}{the bin numbers you wish to export}

\item{obj}{an ODF structure object with multiple depth bins (produced from oce2odf)}

\item{bins}{the bin numbers you wish to export}
}
\description{
Select bin export

Select particular ADCP bins to be exported to ODF files
}
\examples{
```
# oce2odf(adp, write = FALSE)
# bins <- list(12:34)
# binExport(obj = b, bins)
````
```
# oce2odf(adp, write = FALSE)
# bins <- list(12:34)
# binExport(obj = b, bins)
````
}
