% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/read_odf.R
\name{convert_number}
\alias{convert_number}
\alias{convert_number}
\title{convert_number}
\usage{
convert_number(sVal)

convert_number(sVal)
}
\arguments{
\item{sVal}{- a string or array of strings}

\item{sVal}{- a string or array of strings}
}
\description{
Description:
	convert 00000D+00 strings to 00000E+00 strings.

Description:
	convert 00000D+00 strings to 00000E+00 strings.
}
