% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/define_ODF_header.R
\name{define_ODF_header}
\alias{define_ODF_header}
\alias{define_ODF_header}
\title{Defines and returns the header section of an ODF file.
Output:
  ODF_HEADER is a cell array containing the definition.}
\usage{
define_ODF_header()

define_ODF_header()
}
\description{
This tool usually is called by other tools such as ODF read,
  write and edit etc. It is one of the core tools and users 
  should not make any changes to it.

Defines and returns the header section of an ODF file.
Output:
  ODF_HEADER is a cell array containing the definition.
}
\details{
A new block can be added easily to this definition to extend an
  ODF header structure. The new block should be considered as
  'optional', otherwise all existing ODF files will have to be
  modified as well.

Report any bugs to DataServicesDonnees@dfo-mpo.gc.ca

ODSToolbox Version: 2.0

Last Updated: September 3, 2015


Source:
  Ocean Data and Information Services,
  Bedford Institute of Oceanography, DFO, Canada.
  DataServicesDonnees@dfo-mpo.gc.ca

This tool usually is called by other tools such as ODF read,
  write and edit etc. It is one of the core tools and users
  should not make any changes to it.

A new block can be added easily to this definition to extend an
  ODF header structure. The new block should be considered as
  'optional', otherwise all existing ODF files will have to be
  modified as well.

Report any bugs to DataServicesDonnees@dfo-mpo.gc.ca


Source:
  Ocean Data and Information Services,
  Bedford Institute of Oceanography, DFO, Canada.
  DataServicesDonnees@dfo-mpo.gc.ca
}
\examples{
ODF_header <- define_ODF_header()

ODF_header <- define_ODF_header()

}
\author{
Yongcun Hu, Patrick Upson


Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.

Yongcun Hu, Patrick Upson
}
