% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/odfedit.R
\name{editData}
\alias{editData}
\alias{editData}
\title{edit an ODF data frame}
\usage{
editData(odffile)

editData(odffile)
}
\arguments{
\item{odffile}{an odf file connection}

\item{odffile}{an odf file connection}
}
\value{
the data frame from the odf file

the data frame from the odf file
}
\description{
Function returns ODF data frame which can be edited/ processed as necessary then written out again

Function returns ODF data frame which can be edited/ processed as necessary then written out again
}
