% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/odfedit.R
\name{editParam}
\alias{editParam}
\alias{editParam}
\title{Edit an ODF attribute}
\usage{
editParam(odffile, param, value)

editParam(odffile, param, value)
}
\arguments{
\item{odffile}{an ODF file connection}

\item{param}{The parameter you wish to edit}

\item{value}{The value you wish to insert into the odf file}

\item{odffile}{an ODF file connection}

\item{param}{The parameter you wish to edit}

\item{value}{The value you wish to insert into the odf file}
}
\value{
an updated ODF file

an updated ODF file
}
\description{
Edit an ODF attribute

Edit an ODF attribute
}
