% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/oce2odf.R
\name{oce2odfHeader}
\alias{oce2odfHeader}
\alias{oce2odfHeader}
\title{oce2odfHeader}
\usage{
oce2odfHeader(obj)

oce2odfHeader(obj)
}
\arguments{
\item{obj}{oce object}

\item{obj}{oce object}
}
\description{
creates ODF standard header from metadata within an oce object to be copied and replicated in ODF files
object can subsequently be used in function `oce2odf`

creates ODF standard header from metadata within an oce object to be copied and replicated in ODF files
object can subsequently be used in function `oce2odf`
}
\details{
returns empty odf object with header metadata filled out

returns empty odf object with header metadata filled out
}
