% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODF-package.R
\docType{package}
\name{ODF}
\alias{ODF}
\alias{ODF-package}
\title{Ocean Data Format (ODF)}
\description{
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
}
\details{
ODSToolbox Version: 2.0

This program was totally re-designed and re-written for
  ODSToolbox Version 2.0. It is not based on Version 1.0.
  In addition the package, classes, methods and functions,
  originally written for Matlab were, re-written again to
  be used in R.

  While this new version of read_odf corrects many errors in
  Version 1.0, and includes some new functionalities such as
  checking if all mandatory header blocks and all mandatory
  fields are presented in the input ODF file etc., it is possible
  that this program may have some conflicts with other tools in
  ODSToolbox, please email yongcun.hu@dfo-mpo.gc.ca if the
  user find any problems.

See also \code{\link{read_odf}}, \code{\link{write_odf}}.
}
\author{
Yongcun Hu, Patrick Upson
}
