% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odfedit.R
\name{odfSetMetadata}
\alias{odfSetMetadata}
\title{odfSetMetadata}
\usage{
odfSetMetadata(odffile, param, value)
}
\arguments{
\item{odffile}{ODF file connection}

\item{param}{parameter of metadata to be changed}

\item{value}{new value of metadata parameter}
}
\value{
an oce odf object
}
\description{
odfSetMetadata
}
