% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/ParameterReshape.R,
%   R/read_odf.R, R/read_odf_old.R
\name{parameterReshape}
\alias{parameterReshape}
\alias{parameterReshape}
\alias{parameterReshape}
\alias{parameterReshape}
\title{parameterReshape}
\usage{
parameterReshape(paramList)

parameterReshape(paramList)

parameterReshape(paramList)

parameterReshape(paramList)
}
\arguments{
\item{paramList}{The parameter header to be reshaped}

\item{paramList}{The parameter header to be reshaped}

\item{paramList}{The parameter header to be reshaped}
}
\description{
Description:
	Used to convert lists of lists into matrices inorder to present
	the data in a more compact easier to read, and access, way.

Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.

Description:
	Used to convert lists of lists into matrices inorder to present
	the data in a more compact easier to read, and access, way.

Description:
	Used to convert lists of lists into matrices inorder to present
	the data in a more compact easier to read, and access, way.
}
\author{
Patrick Upson
}
