% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/read_odf.R
\name{read_odf}
\alias{read_odf}
\alias{read_odf}
\title{READ_ODF: Read in an ODF file.
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
Description:
  Read in an ODF file.}
\usage{
read_odf(filename)

read_odf(filename)
}
\arguments{
\item{filename}{location and name of the ODF file to be processed}

\item{filename}{location and name of the ODF file to be processed}
}
\description{
READ_ODF: Read in an ODF file.
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
Description:
  Read in an ODF file.

@copyright
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
}
\details{
ODSToolbox Version: 2.0

Last Updated: September 3, 2015

Source:
  Ocean Data and Information Services,
  Bedford Institute of Oceanography, DFO, Canada.
  DataServicesDonnees@dfo-mpo.gc.ca

Notes:
  This program was totally re-designed and re-written for
  ODSToolbox Version 2.0. It is not based on Version 1.0.

  While this new version of read_odf corrects many errors in
  Version 1.0, and includes some new functionalities such as
  checking if all mandatory header blocks and all mandatory
  fields are presented in the input ODF file etc., it is possible
  that this program may have some conflicts with other tools in
  ODSToolbox, please email yongcun.hu@dfo-mpo.gc.ca if the
  user find any problems.

See also \code{\link{write_odf}}.

ODSToolbox Version: 2.0

Last Updated: September 3, 2015

Source:
  Ocean Data and Information Services,
  Bedford Institute of Oceanography, DFO, Canada.
  DataServicesDonnees@dfo-mpo.gc.ca

Notes:
  This program was totally re-designed and re-written for
  ODSToolbox Version 2.0. It is not based on Version 1.0.

  While this new version of read_odf corrects many errors in
  Version 1.0, and includes some new functionalities such as
  checking if all mandatory header blocks and all mandatory
  fields are presented in the input ODF file etc., it is possible
  that this program may have some conflicts with other tools in
  ODSToolbox, please email yongcun.hu@dfo-mpo.gc.ca if the
  user find any problems.

See also \code{\link{write_odf}}.
}
\author{
Yongcun Hu, Patrick Upson

Modified by Gordana Lazin, June 8, 2015
Import data using read.table function (line 190), much faster 
Replace formats for SYTM_01: from strings to ISO time format (asumes UTC), line 231

Yongcun Hu, Patrick Upson

Modified by Gordana Lazin, June 8, 2015
Import data using read.table function (line 190), much faster
Replace formats for SYTM_01: from strings to ISO time format (asumes UTC), line 231
}
