% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/read_odf.R, R/read_odf_old.R
\name{readFile}
\alias{readFile}
\alias{readFile}
\alias{readFile}
\title{readFile}
\usage{
readFile(filename)

readFile(filename)

readFile(filename)
}
\arguments{
\item{filename}{input ODF file name}

\item{filename}{input ODF file name}

\item{filename}{input ODF file name}
}
\description{
Description:
  read the data from a file and return an array of lines

Description:
  read the data from a file and return an array of lines

Description:
  read the data from a file and return an array of lines
}
\details{
Output : OUTPUT = array of file data

Output : OUTPUT = array of file data

Output : OUTPUT = array of file data
}
