% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceDir.R
\name{sourceDir}
\alias{sourceDir}
\title{Source a directory}
\usage{
sourceDir(file_path)
}
\arguments{
\item{file_path}{the path of the directory to source}
}
\description{
Source a directory
}
\author{
: Unknown
Source an entire directory, useful for quickly incorporating a series of .R
function files
}
