% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/write_odf.R, R/write_odf_old.R
\name{write_odf}
\alias{write_odf}
\alias{write_odf}
\alias{write_odf}
\title{write_odf
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
Description:
takes an ODF object and an output location and writes an ODF compliant output file}
\usage{
write_odf(odf_object, output_file)

write_odf(odf_object, output_file)

write_odf(odf_object, output_file)
}
\arguments{
\item{odf_object}{Object to be written to file}

\item{output_file}{location and name of the file the ODF object is to be written to}

\item{odf_object}{Object to be written to file}

\item{output_file}{location and name of the file the ODF object is to be written to}

\item{odf_object}{Object to be written to file}

\item{output_file}{location and name of the file the ODF object is to be written to}
}
\description{
write_odf
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
Description:
takes an ODF object and an output location and writes an ODF compliant output file

write_odf
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
Description:
takes an ODF object and an output location and writes an ODF compliant output file

write_odf
Copyright (C) 2006-2014 DFO, Bedford Institute of Oceanography, Canada.
You may distribute under the terms of either the GNU General Public
License or the Apache v2 License, as specified in the README file.
Description:
takes an ODF object and an output location and writes an ODF compliant output file
}
\details{
date: July 4, 2014

Modified Jun 21, 2018.
#
R.Pettipas

Changed the accessing of PARAMETER_HEADER fields from
#
 odf_object$PARAMETER_HEADER[[field]][i] (which returns NULL)
 to:
 odf_object$PARAMETER_HEADER[i][[1]][[field]]

Changed 
prams <- odf_object$PARAMETER_HEADER$CODE
to:
or (a in 1:length(odf_object$PARAMETER_HEADER)) prams[a] <- B$PARAMETER_HEADER[a][[1]]$CODE


Also added code to change the data type for the "SYTM" parameter to "SYTM" and other values in the parameter header
to properly reflect the data type (if not already set to type "SYTM"). Otherwise, the field is output as a floating point value.

Other changes indicated with comment lines "R.Pettipas"

CAVEAT: If the value of TYPE in the parameter header for an SYTM time field is 'SING' and
not 'SYTM', the output file will have an integer (POSIXct) format for time and not the expected "DD-MMM-YYYY HH:MM:SS.FF" format 
so it is advised to change it before running "write_odf"

date: July 4, 2014

Modified Jun 21, 2018.
#
R.Pettipas

Changed the accessing of PARAMETER_HEADER fields from
#
 odf_object$PARAMETER_HEADER[[field]][i] (which returns NULL)
 to:
 odf_object$PARAMETER_HEADER[i][[1]][[field]]

Changed 
prams <- odf_object$PARAMETER_HEADER$CODE
to:
or (a in 1:length(odf_object$PARAMETER_HEADER)) prams[a] <- B$PARAMETER_HEADER[a][[1]]$CODE


Also added code to change the data type for the "SYTM" parameter to "SYTM" and other values in the parameter header
to properly reflect the data type (if not already set to type "SYTM"). Otherwise, the field is output as a floating point value.

Other changes indicated with comment lines "R.Pettipas"

CAVEAT: If the value of TYPE in the parameter header for an SYTM time field is 'SING' and
not 'SYTM', the output file will have an integer (POSIXct) format for time and not the expected "DD-MMM-YYYY HH:MM:SS.FF" format 
so it is advised to change it before running "write_odf"

date: July 4, 2014
}
\author{
Patrick Upson

Patrick Upson

Patrick Upson
}
