% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/write_odf.R
\name{writeValue}
\alias{writeValue}
\alias{writeValue}
\title{writeValue}
\usage{
writeValue(var_name, var_val, var_head, con)

writeValue(var_name, var_val, var_head, con)
}
\arguments{
\item{var_name}{ODF variable name being written}

\item{var_val}{value associated with the ODF varialbe}

\item{var_head}{ODF header containing information about the ODF variable}

\item{con}{location data is being written to}

\item{var_name}{ODF variable name being written}

\item{var_val}{value associated with the ODF varialbe}

\item{var_head}{ODF header containing information about the ODF variable}

\item{con}{location data is being written to}
}
\description{
Description:
 used to set the type of a variable, format the output string and print
 the string to a file.

Description:
 used to set the type of a variable, format the output string and print
 the string to a file.
}
