% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfspec.R
\name{bfspec}
\alias{bfspec}
\title{@title BFSPEC:}
\usage{
bfspec(C)
}
\description{
Creates a file specification string based on ODF header information.
  ODSToolbox Version: 2.0
}
\details{
Creation Date: 02-SEP-2015
  Last Updated: 02-SEP-2015

@author: Jeff Jackson

@version: 1.0

@copyright: 2014, Fisheries and Oceans Canada. All Rights Reserved.

Source:
      Ocean Data and Information Services,
      Bedford Institute of Oceanography, DFO, Canada.
      DataServicesDonnees"at"dfo-mpo.gc.ca

@summary: Creates a file specification string based on ODF header information.
}
