% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/gen_odfstruct.R,
%   R/gen_odfstruct_old.R
\name{gen_odfstruct}
\alias{gen_odfstruct}
\alias{gen_odfstruct}
\alias{gen_odfstruct}
\title{'''
------------------------------------------------------------------------
GEN_ODFSTRUCT: Build an empty ODF structured array.}
\usage{
gen_odfstruct()

gen_odfstruct()

gen_odfstruct()
}
\description{
ODS Toolbox Version: 2.0

creates an empty list structure to populate and export as ODF

ODS Toolbox Version: 2.0
}
\details{
Created: 19-OCT-2016
Updated: 30-NOV-2016

Created: 19-OCT-2016
Updated: 30-NOV-2016
}
\author{
: Jeff Jackson

: Jeff Jackson

: Jeff Jackson
}
