% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null2na.R
\name{null2na}
\alias{null2na}
\title{null2na}
\usage{
null2na(A)
}
\description{
Replace null values (-99) with NA in the data of an ODF
  structure. Substitutes null data with empty strings in an ODF structure.
}
\details{
ODSToolbox Version: 2.0

Creation Date: 09-SEP-2015
  Last Updated: 12-NOV-2015

@author: Jeff Jackson

@version: 1.0

@copyright: Fisheries and Oceans Canada

Source:
      Ocean Data and Information Services,
      Bedford Institute of Oceanography, DFO, Canada.
      DataServicesDonnees@dfo-mpo.gc.ca

@example

Usage:
      A = null2na(A)

Input:
      A : an ODF structured array

Output:
      A : the updated ODF structured array with null data replaced with NA

Example:
      A = null2na(A)





@Updates:

Jeff Jackson (22-SEP-2015)
    - Fixed an error.

Jeff Jackson (12-NOV-2015)
    - Fixed a logic error.

@bugs
  Report any bugs to DataServicesDonnees@dfo-mpo.gc.ca
}
