% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/oce2odf.R
\name{oce2odf}
\alias{oce2odf}
\alias{oce2odf}
\title{oce2odf}
\usage{
oce2odf(obj, write = TRUE, inst_code = NULL)

oce2odf(obj, write = TRUE, inst_code = NULL)
}
\arguments{
\item{obj}{oce object for data to be copied to odf}

\item{write}{whether or not to write out all the odf files produced, default
is TRUE, if false please use binExport to select the bins for which you
would like to produce ODFs}

\item{inst_code}{instrument code acronym for file name}

\item{obj}{oce object for data to be copied to odf}

\item{write}{whether or not to write out all the odf files produced, default
is TRUE, if false please use binExport to select the bins for which you
would like to produce ODFs}
}
\description{
Produces ODF files from ADCP data set (currently only accepts oce adp object,
ideally will accept various oce instrument objects)

Currently fully functioning for ADCP data
Accepts an oce class object (adp) and converts into a series of ODF files to
be exported
}
\details{
Can also accept other moored instruments and convert from oce objects to odf
including mctd, mtr and mcm


In order to function requires certain metadata items
* depthMean
* time_coverage_start
* time_coverage_end
* depthMin
* depthMax
* event_comments
* serialNumber
* mooring_number
* cruise_number

Examples of instrument codes: MADCPS (Moored ADCP), MCTD (moored CTD), MTR
(Moored temperature recorder), MCM (moored current meter)
}
\author{
E. Chisholm
Version 1.0
}
