% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODFtoolbox.R, R/oce2odf.R
\name{oce2odf}
\alias{oce2odf}
\alias{oce2odf}
\title{oce2odf}
\usage{
oce2odf(obj, write = TRUE)

oce2odf(obj, write = TRUE)
}
\arguments{
\item{obj}{oce object for data to be copied to odf}

\item{write}{whether or not to write out all the odf files produced, default
is TRUE, if false please use binExport to select the bins for which you
would like to produce ODFs}

\item{obj}{oce object for data to be copied to odf}

\item{write}{whether or not to write out all the odf files produced, default
  is TRUE, if false please use binExport to select the bins for which you
  would like to produce ODFs

  @export}
}
\description{
Produces ODF files from ADCP data set (currently only accepts oce adp object,
ideally will accept various oce instrument objects)

Currently fully functioning for ADCP data
Accepts an oce class object (adp) and converts into a series of ODF files to
be exported
}
